Add files via upload
HistGradientBoostingClassifier provides a powerful mechanism for classifying and distinguishing between different types of anomalies. The model is designed to detect and classify anomalies within a dataset, making it a valuable tool for identifying unusual or suspicious patterns in various applications, including cybersecurity, fraud detection, and quality control. The model's effectiveness and efficiency make it a robust solution for real-world anomaly detection challenges.